Revert "Move commands to a shared location"
This reverts commit a005fc1d4f6cf9b8075b04338f5b71280ac0015a.
